[BugFix] Comparison report crashes for new periods in cpass0 as nothing to compare to

+++ TPCQCVis/tools/generateReport.py
@@ -111,31 +111,34 @@ if __name__ == "__main__":
                 os.remove(temp_period_path)
 
     #### Comparison Report #####
-    # Create a temporary file with a unique filename for the period report
-    with tempfile.NamedTemporaryFile(prefix="TPCQC_", suffix=".ipynb", delete=False) as temp_comparison:
-        temp_comparison_path = temp_comparison.name
+    if len(glob.glob(args.path + "/" + args.period + "/*")) > 1:
+        # Create a temporary file with a unique filename for the period report
+        with tempfile.NamedTemporaryFile(prefix="TPCQC_", suffix=".ipynb", delete=False) as temp_comparison:
+            temp_comparison_path = temp_comparison.name
 
-    replace_in_ipynb(comparison_template_path, temp_comparison_path,
-        ["myPeriod", "myPass", "myPath"],
-        [ args.period, args.apass, args.path]
-    )
+        replace_in_ipynb(comparison_template_path, temp_comparison_path,
+            ["myPeriod", "myPass", "myPath"],
+            [ args.period, args.apass, args.path]
+        )
 
-    # The command and its arguments for the comparison report
-    comparison_report_command = [
-        "jupyter", "nbconvert", temp_comparison_path, "--to", "html", "--template", "classic", "--no-input", "--execute",
-        "--output", args.path + "/" + args.period + "/" + args.period + "_" + args.apass +"_comparison.html"
-    ]
+        # The command and its arguments for the comparison report
+        comparison_report_command = [
+            "jupyter", "nbconvert", temp_comparison_path, "--to", "html", "--template", "classic", "--no-input", "--execute",
+            "--output", args.path + "/" + args.period + "/" + args.period + "_" + args.apass +"_comparison.html"
+        ]
 
-    # Run the command for the comparison report
-    output = subprocess.run(comparison_report_command, capture_output=True)
+        # Run the command for the comparison report
+        output = subprocess.run(comparison_report_command, capture_output=True)
 
-    # Check the return code of the command
-    if output.returncode == 0:
-        # If the command runs successfully
-        print("comparison QC report generated successfully for period", args.period)
+        # Check the return code of the command
+        if output.returncode == 0:
+            # If the command runs successfully
+            print("comparison QC report generated successfully for period", args.period)
+        else:
+            # If the command fails
+            print("Error:", output.stderr.decode())
+
+        if temp_comparison_path:
+            os.remove(temp_comparison_path)
     else:
-        # If the command fails
-        print("Error:", output.stderr.decode())
-
-    if temp_comparison_path:
-        os.remove(temp_comparison_path)
+        print("Not running comparison report, as not enough periods!")
